Remove legacy Jinja2 templates and view functions
