--- tahrir/templates/about.html
@@ -1,13 +1,0 @@
-{% extends '_base.html' %}
-
-{% block body %}
-
-<div class="page">
-  <div class="grid-50 push-25">
-    <div class="shadow">
-      {{content | safe }}
-    </div>
-  </div>
-</div>
-
-{% endblock %}
