--- tahrir/templates/explore_badges.html
@@ -1,41 +1,0 @@
-{% from '_functions.html' import badge_thumbnail_flex, show_badges_by_tags with context %}
-{% extends '_base.html' %}
-
-
-{% block head %}
-<link rel="alternate" type="application/rss+xml" title="Newest badges" href="{{ url_for("tahrir.explore_badges_rss") }}" />
-{% endblock %}
-
-
-{% block body %}
-
-<div class="page">
-
-  <div class="clear spacer"></div>
-
-
-
-  <!-- COLUMN 1 (Left)-->
-  <div class="grid-50">
-    <div class="shadow">
-      <h1 class="section-header">All Badges</h1>
-      <div class="padded-content">
-        {{ show_badges_by_tags(all_badges_by_tag, all_uncategorized)}}
-      </div> <!-- End padded content. -->
-    </div> <!-- End shadow. -->
-  </div>
-
-  <div class="grid-50">
-    <div class="shadow">
-      <h1 class="section-header">Newest Badges</h1>
-      <div class="padded-content">
-      {{ show_badges_by_tags(newest_badges_by_tag, newest_uncategorized)}}
-      </div> <!-- End padded content. -->
-    </div> <!-- End shadow. -->
-  </div> <!-- End grid-50 -->
-
-  <div class="clear spacer"></div>
-</div>
-
-
-{% endblock %}
