--- tahrir/templates/tags.html
@@ -1,32 +1,0 @@
-{% extends '_base.html' %}
-
-{% block body %}
-
-<div class="page">
-  <div class="grid-100">
-    <div class="shadow">
-      <h1 class="section-header">Tag Info</h1>
-      <div class="padded-content">
-        <p>
-          You searched for
-          {% for t in tags %}
-            {% if loop.last %}
-              <strong>{{t}}</strong>
-            {% else %}
-              {{t}},
-            {% endif %}
-          {% endfor %}
-        </p>
-        <ul class="pretty-list">
-          {% for badge in badges %}
-            <li><a href="{{ url_for('tahrir.badge', badge_id=badge.id) }}">{{badge.name}}</a></li>
-          {% endfor %}
-        </ul>
-      </div> <!-- End shadow. -->
-    </div> <!-- End padded content. -->
-
-  </div> <!-- End grid-100. -->
-  <div class="clear spacer"></div>
-</div>
-
-{% endblock %}
