+++ tahrir/views/__init__.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, Flask, render_template
+from flask import Blueprint, Flask, jsonify
 from flask.helpers import send_from_directory
 
 
@@ -6,29 +6,11 @@ blueprint = Blueprint("tahrir", __name__)
 
 
 def page_not_found(e):
-    title = "Oops! The page you are looking for doesn't exist."
-    message = (
-        "It might have been removed or had its name changed. "
-        "Please check the URL or return to the home page."
-    )
-    return (
-        render_template("error.html", error_code=404, error_title=title, error_message=message),
-        404,
-    )
+    return jsonify({"error": "Resource not found", "status": 404}), 404
 
 
 def internal_server_error(e):
-    title = "Whoops! There's badness in the server internals."
-    message = (
-        "Please try refreshing the page or come back later. "
-        "If the issue persists, please open an "
-        '<a href="https://pagure.io/fedora-infrastructure/new_issue">'
-        "Infrastructure ticket<a>."
-    )
-    return (
-        render_template("error.html", error_code=500, error_title=title, error_message=message),
-        500,
-    )
+    return jsonify({"error": "Internal server error", "status": 500}), 500
 
 
 def add_static_view(app: Flask, directory: str, prefix: str, endpoint: str):
